hw3 (Переделано)
