+++ src/Homework3/ScenarioMain.java
@@ -1,0 +1,22 @@
+package Homework3;
+
+public class ScenarioMain extends Scenario{
+
+    public static void main(String[] args) {
+        String[] roles = {
+                "Герцог",
+                "Граф",
+                "Графиня",
+                "Гость"};
+        String[] textLines = {"Графиня: Кто вы?!",
+                "Гость: Я… Я тот, которого вы видите перед собой.",
+                "Графиня: Очень остроумно. Что-нибудь случилось?",
+                "Гость: Ничего особенного, но… Дело в том, что… Позвольте сказать вам пару слов по секрету.",
+                "Графиня: Вы?",
+                "Гость: Так уж получилось",
+                "Графиня: Но почему в таком виде?",
+                "Гость: Здесь не время и не место объяснять это.",
+                "Графиня: Вы правы.Садитесь, дорогой герцог"};
+        System.out.println(printTextPerRole(roles, textLines));
+    }
+}
